requirements.txt
transformers
torch
datasets
bitsandbytes
wandb
accelerate
bitsandbytes
numpy
evaluate
peft
scikit-learn
click
pyyaml
ipdb
